+++ packages/solid/src/methods/clearError.ts
@@ -1,0 +1,16 @@
+import { FieldArrayPath, FieldPath, FieldValues, FormState } from '../types';
+import { setError } from './setError';
+
+/**
+ * Clears the error of the specified field or field array.
+ *
+ * @param form The form that contains the field.
+ * @param name The name of the field where the error should be cleared.
+ */
+export function clearError<
+  TFieldValues extends FieldValues,
+  TFieldName extends FieldPath<TFieldValues>,
+  TFieldArrayName extends FieldArrayPath<TFieldValues>
+>(form: FormState<TFieldValues>, name: TFieldName | TFieldArrayName): void {
+  setError(form, name, '');
+}